+++ R/RcppExports.R
@@ -29,8 +29,8 @@ get_vals_lagcrit <- function(y, x, lag_crit, h, k, max_lags, n_obs) {
 #' @name hp_filter
 #' @title Decompose a times series via the Hodrick-Prescott filter
 #' @description  Estimate cyclical and trend component with filter by Hodrick and Prescott (1997).
-#' The function is based on the function \emph{hpfilter} from the archived package \emph{mFilter}.
-#' @param x Column matrix with numeric values.
+#' The function is based on the function \emph{hpfilter} from the archived \emph{mFilter}-package.
+#' @param x One column matrix with numeric values.
 #' @param lambda Numeric value.
 #' @return A list. The first element contains the cyclical component and the second element the trend component.
 #' @examples
@@ -39,7 +39,7 @@ get_vals_lagcrit <- function(y, x, lag_crit, h, k, max_lags, n_obs) {
 #'
 #'
 #'# Decompose the Federal Funds Rate
-#'  data_set     <- as.matrix(interest_rules_var_data$FF) # Input has to be a one column matrix
+#'  data_set     <- as.matrix(interest_rules_var_data$FF)
 #'  hp_results   <- hp_filter(data_set, 1600)
 #'
 #'# Extract results and save as data.frame
@@ -73,6 +73,25 @@ get_vals_lagcrit <- function(y, x, lag_crit, h, k, max_lags, n_obs) {
 #' @author Philipp Adämmer
 hp_filter <- function(x, lambda) {
     .Call(`_lpirfs_hp_filter`, x, lambda)
+}
+
+#' @name newey_west_tsls
+#' @title Compute 2SLS parameters and robust standard errors based on Newey-West
+#' @description  Compute 2SLS parameters and robust standard errors based on Newey and West (1987).
+#' Part of the function is based on the Matlab code by James P. LeSage.
+#' @param y Numeric vector.
+#' @param x Numeric matrix.
+#' @param z Numeric matrix.
+#' @param h Integer.
+#' @return A list. The first element contains the estimated 2SLS parameters and the second element
+#' the covariance matrix of these parameters.
+#' @keywords internal
+#' @references
+#' Newey, W.K., and West, K.D. (1987). “A Simple, Positive-Definite, Heteroskedasticity and
+#' Autocorrelation Consistent Covariance Matrix.” \emph{Econometrica}, 55, 703–708.
+#' Wooldridge, J.M. (2002), Econometric Analysis of Cross Section and Panel Data, The MIT Press.
+newey_west_tsls <- function(y, x, z, h) {
+    .Call(`_lpirfs_newey_west_tsls`, y, x, z, h)
 }
 
 #' @name newey_west